requirements.txt
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
distlib==0.3.3
entrypoints==0.3
factory-boy==3.2.1
Faker==9.9.0
Flask==2.0.2
greenlet==0.4.17
gunicorn==20.1.0
iniconfig==1.1.1
Jinja2==3.0.3
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
more-itertools==8.12.0
mysql==0.0.3
mysqlclient==2.1.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
SQLAlchemy==1.4.27
text-unidecode==1.3
urllib3==1.25.11
virtualenv==20.10.0
websocket-client==1.2.1
Werkzeug==2.0.2
wrapt==1.13.3
